cache_failures: cache number lookup failures
when an offnet route_req comes in for an unknown number, many whapps
will attempt to resolve the account-id of the dialed number. Since the
number doesn't exist, this will cause lots of 404 requests from
Couch (and wasted time waiting for the 404). By caching the failure, we
can avoid the lookup after the first failure. The failure will be
flushed once the doc id created/modified/deleted.