Logging updates for version mismatches and missing files (#4676)
* log the lack of a file loudly

This particular issue arose on a system that had been upgraded from
4.2.14 to 4.2.15 (meaning .14 paths were gone) but ecallmgr wasn't
restarted and so was still thinking .14 paths were where to find
things.

* log a warning when nodes of the same name have mismatched versions

Since status checks are common, it can be helpful to see if/when a
node is running a different version from the rest. In this case, we
use the 'name' portion of the node name to determine what nodes should
match versions. So kazoo_apps, ecallmgr, kamailio, all will compare
versions properly.